Final-Hackathon/Backend/requirements.txt
langgraph>=0.1.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
whisper>=1.1.10
pydub>=0.25.1
librosa>=0.10.1
yt-dlp>=2023.12.30
numpy>=1.24.0
fastapi
uvicorn
pydantic
motor